requirements.txt
appdirs==1.4.4
autopep8==1.5.3
black==21.7b0
click==8.0.1
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
pycodestyle==2.6.0
python-dotenv==0.13.0
regex==2021.8.3
six==1.15.0
toml==0.10.1
tomli==1.2.1
virtualenv==20.0.27
Werkzeug==2.0.2
